terminal.c: Also set the attribs on the 2nd part of wide chars
In do_paint(), if newline[j] is a wide char (tattr & ATTR_WIDE), putty
steps the both j, and the pointer (d) which points to the internal state's
character (lchars[j]), and copies its value to term->disptext[i]->chars[j],
but did not copy the attributes with it. With this patch both parts of the
wide character will have ATTR_WIDE, helping the "frontend" to draw it
properly.

References: jira PAM-8611